+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/doc
 
+## 0.15.14
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.15.14
+
 ## 0.15.13
 
 ### Patch Changes